+++ src/Concordium/Client/Types/Contract/Parameter.hs
@@ -6,6 +6,7 @@ module Concordium.Client.Types.Contract.Parameter where
 import Concordium.Client.Config (showCompactPrettyJSON, showPrettyJSON)
 import qualified Concordium.Types as T
 import Concordium.Client.Types.Contract.Schema
+import qualified Concordium.Wasm as Wasm
 import qualified Data.DoubleWord as DW
 
 import Control.Monad (unless, when, zipWithM)
@@ -79,13 +80,22 @@ getJSONUsingSchema typ = case typ of
                       Nothing -> fail [i|Variant with index #{idx} does not exist for Enum.|]
     fields' <- getFieldsAsJSON fields
     return $ AE.object [name .= fields']
-  String sl -> do
-    bStr <- BS.pack <$> getListOfWithSizeLen sl S.get
-    case Text.decodeUtf8' bStr of
-      Left _ -> fail "String is not valid UTF-8."
-      Right str -> return $ AE.toJSON str
+  String sl -> AE.toJSON <$> getUtf8String sl
   UInt128 -> AE.toJSON . show <$> DW.getWord128le
   Int128 -> AE.toJSON . show <$> DW.getInt128le
+  ContractName sl -> do
+    rawName <- getUtf8String sl
+    case Text.stripPrefix "init_" rawName of
+      Just contractName -> return $ AE.object ["contract" .= contractName]
+      Nothing -> fail [i|Expect contract name with format: 'init_<contract_name>', but got: #{rawName}.|]
+  ReceiveName sl -> do
+    rawName <- getUtf8String sl
+    let separatorIndex = Text.findIndex (== '.') rawName
+    case separatorIndex of
+      Nothing -> fail [i|Expected format '<contract_name>.<func_name>' but got: #{rawName}.|]
+      Just idx -> let (contractName, funcNameWithDot) = Text.splitAt idx rawName
+                      funcName = Text.tail funcNameWithDot -- Safe, since we know it contains a dot.
+                  in return $ AE.object ["contract" .= contractName, "func" .= funcName]
 
   where
     getFieldsAsJSON :: Fields -> S.Get AE.Value
@@ -122,6 +132,14 @@ getJSONUsingSchema typ = case typ of
             if value == 0
             then Nothing
             else Just [i|#{value}#{unit}|]
+
+    -- |Gets a string of the specified SizeLength. Fails if the string is not valid UTF8.
+    getUtf8String :: SizeLength -> S.Get Text
+    getUtf8String sl = do
+      bStr <- BS.pack <$> getListOfWithSizeLen sl S.get
+      case Text.decodeUtf8' bStr of
+        Left _ -> fail "String is not valid UTF-8."
+        Right str -> return str
 
 -- |Create a `Serialize.Put` for JSON using a `SchemaType`.
 -- It goes through the JSON and SchemaType recursively, and
@@ -222,6 +240,42 @@ putJSONUsingSchema typ json = case (typ, json) of
     case DW.int128FromString $ Text.unpack str of
       Left err -> Left [i|Invalid Int128 '#{str}': #{err}|]
       Right n -> pure $ DW.putInt128le n
+
+  (ContractName sl, AE.Object obj) -> do
+    let fieldCount = length . HM.toList $ obj
+    when (fieldCount /= 1) $ Left [i|Expected object with a single field 'contract', but got: #{showPrettyJSON obj}.|]
+    case HM.lookup "contract" obj of
+      Just (AE.String contractName) -> do
+        let nameWithInit = "init_" <> contractName
+            bytes = BS.unpack . Text.encodeUtf8 $ nameWithInit
+            len = fromIntegral $ length bytes
+            maxLen = maxSizeLen sl
+        unless (Wasm.isValidInitName nameWithInit) $
+          -- Include 'obj' in the error to be explicit about where the error occured.
+          Left [i|'#{contractName}' is not a valid contract name.\nIn #{showPrettyJSON obj}.|]
+        when (len > maxLen) $ Left $ tooLongError "ContractName" maxLen len
+        let putLen = putLenWithSizeLen sl $ fromIntegral len
+        let putBytes = mapM_ S.put bytes
+        pure $ putLen <> putBytes
+      _ -> Left [i|Expected object with a field 'contract' of type String, but got: #{showPrettyJSON obj}.|]
+
+  (ReceiveName sl, AE.Object obj) -> do
+    let fieldCount = length . HM.toList $ obj
+    when (fieldCount /= 2) $ Left [i|Expected object with two fields 'contract' and 'func', but got: #{showPrettyJSON obj}.|]
+    case (HM.lookup "contract" obj, HM.lookup "func" obj) of
+      (Just (AE.String contractName), Just (AE.String funcName)) -> do
+        let nameWithDot = contractName <> "." <> funcName
+            bytes = BS.unpack . Text.encodeUtf8 $ nameWithDot
+            len = fromIntegral $ length bytes
+            maxLen = maxSizeLen sl
+        unless (Wasm.isValidReceiveName nameWithDot) $
+          -- Include 'obj' in the error to be explicit about where the error occured.
+          Left [i|'#{contractName}' combined with '#{funcName}' is not a valid receive name.\nIn #{showPrettyJSON obj}.|]
+        when (len > maxLen) $ Left $ tooLongError "ReceiveName" maxLen len
+        let putLen = putLenWithSizeLen sl $ fromIntegral len
+        let putBytes = mapM_ S.put bytes
+        pure $ putLen <> putBytes
+      _ -> Left [i|Expected object with fields 'contract' and 'func' of type String, but got: #{showPrettyJSON obj}.|]
 
   (type_, value) -> Left [i|Expected value of type #{showCompactPrettyJSON type_}, but got: #{showCompactPrettyJSON value}.|]