+++ ChangeLog.md
@@ -1,4 +1,5 @@
 # Changelog for concordium-client
 
 ## Unreleased changes
+- Add ContractName and ReceiveName to schema.
 - Add UInt128 and Int128 to schema type